+++ modules/apps/apio-architect/apio-architect-api/src/main/java/com/liferay/apio/architect/language/Language.java
@@ -22,9 +22,12 @@ import java.util.stream.Stream;
 /**
  * Represents information about the requested languages.
  *
- * @author Alejandro Hernández
+ * @author     Alejandro Hernández
+ * @deprecated As of 1.0.0, should use {@code AcceptLanguage}
+ * @review
  */
 @ConsumerType
+@Deprecated
 @FunctionalInterface
 public interface Language {