+++ modules/apps/asset/asset-publisher-web/src/main/java/com/liferay/asset/publisher/web/internal/exportimport/portlet/preferences/processor/AssetPublisherExportImportPortletPreferencesProcessor.java
@@ -131,11 +131,15 @@ public class AssetPublisherExportImportPortletPreferencesProcessor
 		throws PortletDataException {
 
 		try {
-			if (MapUtil.getBoolean(
-					portletDataContext.getParameterMap(),
-					PortletDataHandlerKeys.PORTLET_DATA) &&
-				!MergeLayoutPrototypesThreadLocal.isInProgress()) {
+			if (MergeLayoutPrototypesThreadLocal.isInProgress()) {
+				if (MapUtil.getBoolean(
+						portletDataContext.getParameterMap(),
+						PortletDataHandlerKeys.PORTLET_DATA)) {
 
+					exportAssetObjects(portletDataContext, portletPreferences);
+				}
+			}
+			else {
 				exportAssetObjects(portletDataContext, portletPreferences);
 			}
 
@@ -320,9 +324,8 @@ public class AssetPublisherExportImportPortletPreferencesProcessor
 				uuid = assetCategory.getUuid();
 				groupId = assetCategory.getGroupId();
 
-				portletDataContext.addReferenceElement(
-					portlet, rootElement, assetCategory,
-					PortletDataContext.REFERENCE_TYPE_DEPENDENCY, true);
+				StagedModelDataHandlerUtil.exportReferenceStagedModel(
+					portletDataContext, portlet.getPortletId(), assetCategory);
 			}
 		}
 		else if (className.equals(AssetVocabulary.class.getName())) {
@@ -334,9 +337,9 @@ public class AssetPublisherExportImportPortletPreferencesProcessor
 				uuid = assetVocabulary.getUuid();
 				groupId = assetVocabulary.getGroupId();
 
-				portletDataContext.addReferenceElement(
-					portlet, rootElement, assetVocabulary,
-					PortletDataContext.REFERENCE_TYPE_DEPENDENCY, true);
+				StagedModelDataHandlerUtil.exportReferenceStagedModel(
+					portletDataContext, portlet.getPortletId(),
+					assetVocabulary);
 			}
 		}
 		else if (className.equals(DDMStructure.class.getName())) {
@@ -347,9 +350,8 @@ public class AssetPublisherExportImportPortletPreferencesProcessor
 				uuid = ddmStructure.getUuid();
 				groupId = ddmStructure.getGroupId();
 
-				portletDataContext.addReferenceElement(
-					portlet, rootElement, ddmStructure,
-					PortletDataContext.REFERENCE_TYPE_DEPENDENCY, true);
+				StagedModelDataHandlerUtil.exportReferenceStagedModel(
+					portletDataContext, portlet.getPortletId(), ddmStructure);
 			}
 		}
 		else if (className.equals(DLFileEntryType.class.getName())) {
@@ -360,9 +362,9 @@ public class AssetPublisherExportImportPortletPreferencesProcessor
 				uuid = dlFileEntryType.getUuid();
 				groupId = dlFileEntryType.getGroupId();
 
-				portletDataContext.addReferenceElement(
-					portlet, rootElement, dlFileEntryType,
-					PortletDataContext.REFERENCE_TYPE_DEPENDENCY, true);
+				StagedModelDataHandlerUtil.exportReferenceStagedModel(
+					portletDataContext, portlet.getPortletId(),
+					dlFileEntryType);
 			}
 		}
 		else if (className.equals(Organization.class.getName())) {
@@ -436,8 +438,8 @@ public class AssetPublisherExportImportPortletPreferencesProcessor
 		}
 		else if (className.equals(DDMStructure.class.getName())) {
 			DDMStructure ddmStructure =
-				_ddmStructureLocalService.fetchDDMStructureByUuidAndGroupId(
-					uuid, groupId);
+				_ddmStructureLocalService.fetchStructureByUuidAndGroupId(
+					uuid, groupId, true);
 
 			if (ddmStructure == null) {
 				Map<String, String> structureUuids =