+++ modules/apps/journal/journal-web/src/main/java/com/liferay/journal/web/internal/display/context/EditArticleDisplayPageDisplayContext.java
@@ -110,9 +110,9 @@ public class EditArticleDisplayPageDisplayContext {
 		}
 
 		AssetDisplayPageEntry assetDisplayPageEntry =
-			AssetDisplayPageEntryLocalServiceUtil.
-				fetchAssetDisplayPageEntryByAssetEntryId(
-					assetEntry.getEntryId());
+			AssetDisplayPageEntryLocalServiceUtil.fetchAssetDisplayPageEntry(
+				assetEntry.getGroupId(), assetEntry.getClassNameId(),
+				assetEntry.getClassPK());
 
 		if (assetDisplayPageEntry != null) {
 			_assetDisplayPageEntry = assetDisplayPageEntry;
@@ -158,6 +158,12 @@ public class EditArticleDisplayPageDisplayContext {
 			getAssetDisplayPageEntry();
 
 		int displayPageType = AssetDisplayPageConstants.TYPE_NONE;
+
+		if (assetDisplayPageEntry == null) {
+			_displayPageType = displayPageType;
+
+			return _displayPageType;
+		}
 
 		if (Objects.equals(
 				assetDisplayPageEntry.getType(),