+++ portal-impl/src/com/liferay/portal/model/impl/PortletImpl.java
@@ -63,6 +63,7 @@ import com.liferay.portal.kernel.template.TemplateHandler;
 import com.liferay.portal.kernel.trash.TrashHandler;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.ContentTypes;
+import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.ServiceProxyFactory;
 import com.liferay.portal.kernel.util.StringBundler;
@@ -336,9 +337,9 @@ public class PortletImpl extends PortletBaseImpl {
 	}
 
 	/**
-	 * Adds a portlet CSS/JS resource dependency.
+	 * Adds a portlet CSS/JavaScript resource dependency.
 	 *
-	 * @param portletDependency a portlet CSS/JS resource dependency
+	 * @param portletDependency the portlet CSS/JavaScript resource dependency
 	 */
 	@Override
 	public void addPortletDependency(PortletDependency portletDependency) {
@@ -859,7 +860,7 @@ public class PortletImpl extends PortletBaseImpl {
 	 */
 	@Override
 	public String getDisplayName() {
-		return _displayName;
+		return GetterUtil.getString(_displayName);
 	}
 
 	/**
@@ -876,7 +877,7 @@ public class PortletImpl extends PortletBaseImpl {
 	 * Returns the Facebook integration method of the portlet.
 	 *
 	 * @return     the Facebook integration method of the portlet
-	 * @deprecated As of 7.0.0, with no direct replacement
+	 * @deprecated As of Judson (7.1.x), with no direct replacement
 	 */
 	@Deprecated
 	@Override
@@ -1442,9 +1443,9 @@ public class PortletImpl extends PortletBaseImpl {
 	}
 
 	/**
-	 * Returns the list of portlet CSS/JS resource dependencies.
+	 * Returns the portlet's CSS/JavaScript resource dependencies.
 	 *
-	 * @return the list of portlet CSS/JS resource dependencies
+	 * @return the portlet's CSS/JavaScript resource dependencies
 	 */
 	@Override
 	public List<PortletDependency> getPortletDependencies() {
@@ -2631,14 +2632,14 @@ public class PortletImpl extends PortletBaseImpl {
 
 	/**
 	 * Returns <code>true</code> if the portlet's
-	 * serveResource(ResourceRequest,ResourceResponse) method should be invoked
-	 * during a partial action triggered by a different portlet on the same
-	 * portal page.
+	 * <code>serveResource(ResourceRequest,ResourceResponse)</code> method
+	 * should be invoked during a partial action triggered by a different
+	 * portlet on the same portal page.
 	 *
 	 * @return <code>true</code> if the portlet's
-	 *         serveResource(ResourceRequest,ResourceResponse) method should be
-	 *         invoked during a partial action triggered by a different portlet
-	 *         on the same portal page
+	 *         <code>serveResource(ResourceRequest,ResourceResponse)</code>
+	 *         method should be invoked during a partial action triggered by a
+	 *         different portlet on the same portal page
 	 */
 	public boolean isPartialActionServeResource() {
 		return _partialActionServeResource;
@@ -2657,36 +2658,29 @@ public class PortletImpl extends PortletBaseImpl {
 	}
 
 	/**
-	 * Returns <code>true</code> if the CSS resource dependencies that are
-	 * specified in portlet.xml, @{@link javax.portlet.annotations.Dependency},
+	 * Returns <code>true</code> if the CSS resource dependencies specified in
+	 * <code>portlet.xml</code>, @{@link javax.portlet.annotations.Dependency},
 	 * {@link javax.portlet.HeaderResponse#addDependency(String, String,
 	 * String)}, or {@link javax.portlet.HeaderResponse#addDependency(String,
 	 * String, String, String)} are to be referenced in the page's header.
 	 *
-	 * @return <code>true</code> if the CSS resource dependencies that are
-	 * specified in portlet.xml, @{@link javax.portlet.annotations.Dependency},
-	 * {@link javax.portlet.HeaderResponse#addDependency(String, String,
-	 * String)}, or {@link javax.portlet.HeaderResponse#addDependency(String,
-	 * String, String, String)} are to be referenced in the page's header
+	 * @return <code>true</code> if the specified CSS resource dependencies are
+	 *         to be referenced in the page's header
 	 */
 	public boolean isPortletDependencyCssEnabled() {
 		return _portletDependencyCssEnabled;
 	}
 
 	/**
-	 * Returns <code>true</code> if the JavaScript resource dependencies that
-	 * are specified in portlet.xml,
-	 * @{@link javax.portlet.annotations.Dependency},
-	 * {@link javax.portlet.HeaderResponse#addDependency(String, String,
-	 * String)}, or {@link javax.portlet.HeaderResponse#addDependency(String,
-	 * String, String, String)} are to be referenced in the page's header.
+	 * Returns <code>true</code> if the JavaScript resource dependencies
+	 * specified in <code>portlet.xml</code>, @{@link
+	 * javax.portlet.annotations.Dependency}, {@link
+	 * javax.portlet.HeaderResponse#addDependency(String, String, String)}, or
+	 * {@link javax.portlet.HeaderResponse#addDependency(String, String, String,
+	 * String)} are to be referenced in the page's header.
 	 *
-	 * @return <code>true</code> if the JavaScript resource dependencies that
-	 * are specified in portlet.xml,
-	 * @{@link javax.portlet.annotations.Dependency},
-	 * {@link javax.portlet.HeaderResponse#addDependency(String, String,
-	 * String)}, or {@link javax.portlet.HeaderResponse#addDependency(String,
-	 * String, String, String)} are to be referenced in the page's header
+	 * @return <code>true</code> if the specified JavaScript resource
+	 *         dependencies are to be referenced in the page's header
 	 */
 	public boolean isPortletDependencyJavaScriptEnabled() {
 		return _portletDependencyJavaScriptEnabled;
@@ -3171,7 +3165,7 @@ public class PortletImpl extends PortletBaseImpl {
 	 *
 	 * @param      facebookIntegration the Facebook integration method of the
 	 *             portlet
-	 * @deprecated As of 7.0.0, with no direct replacement
+	 * @deprecated As of Judson (7.1.x), with no direct replacement
 	 */
 	@Deprecated
 	@Override
@@ -3500,15 +3494,13 @@ public class PortletImpl extends PortletBaseImpl {
 	}
 
 	/**
-	 * Set to <code>true</code> if the portlet's
-	 * serveResource(ResourceRequest,ResourceResponse) method should be invoked
-	 * during a partial action triggered by a different portlet on the same
-	 * portal page.
+	 * Sets whether the portlet's serve resource should be invoked during a
+	 * partial action triggered by a different portlet on the same portal page.
 	 *
-	 * @param partialActionServeResource boolean value for whether the portlet's
-	 *        serveResource(ResourceRequest,ResourceResponse) method should be
-	 *        invoked during a partial action triggered by a different portlet
-	 *        on the same portal page
+	 * @param partialActionServeResource whether the portlet's
+	 *        <code>serveResource(ResourceRequest,ResourceResponse)</code>
+	 *        method should be invoked during a partial action triggered by a
+	 *        different portlet on the same portal page
 	 */
 	@Override
 	public void setPartialActionServeResource(
@@ -3558,11 +3550,11 @@ public class PortletImpl extends PortletBaseImpl {
 	}
 
 	/**
-	 * Set to <code>true</code> if the portlet goes into the pop up state when
-	 * the user goes into the print mode.
+	 * Sets whether the portlet goes into the pop up state when the user goes
+	 * into the print mode.
 	 *
-	 * @param popUpPrint boolean value for whether the portlet goes into the pop
-	 *        up state when the user goes into the print mode
+	 * @param popUpPrint whether the portlet goes into the pop up state when the
+	 *        user goes into the print mode
 	 */
 	@Override
 	public void setPopUpPrint(boolean popUpPrint) {
@@ -3603,19 +3595,14 @@ public class PortletImpl extends PortletBaseImpl {
 	}
 
 	/**
-	 * Set to <code>true</code> if the CSS resource dependencies that are
-	 * specified in portlet.xml, @{@link javax.portlet.annotations.Dependency},
+	 * Sets whether the CSS resource dependencies specified in
+	 * <code>portlet.xml</code>, @{@link javax.portlet.annotations.Dependency},
 	 * {@link javax.portlet.HeaderResponse#addDependency(String, String,
 	 * String)}, or {@link javax.portlet.HeaderResponse#addDependency(String,
 	 * String, String, String)} are to be referenced in the page's header.
 	 *
-	 * @param portletDependencyCssEnabled boolean value for whether the CSS
-	 *        resource dependencies that are specified in portlet.xml,
-	 *        @{@link javax.portlet.annotations.Dependency},
-	 *        {@link javax.portlet.HeaderResponse#addDependency(String, String,
-	 *        String)}, or {@link javax.portlet.HeaderResponse#addDependency(
-	 *        String, String, String, String)} are to be referenced in the
-	 *        page's header
+	 * @param portletDependencyCssEnabled whether the CSS resource dependencies
+	 *        that are specified in <code>portlet.xml</code>,
 	 */
 	public void setPortletDependencyCssEnabled(
 		boolean portletDependencyCssEnabled) {
@@ -3624,19 +3611,19 @@ public class PortletImpl extends PortletBaseImpl {
 	}
 
 	/**
-	 * Set to <code>true</code> if the JavaScript resource dependencies that are
-	 * specified in portlet.xml, @{@link javax.portlet.annotations.Dependency},
+	 * Sets whether the JavaScript resource dependencies specified in
+	 * <code>portlet.xml</code>, @{@link javax.portlet.annotations.Dependency},
 	 * {@link javax.portlet.HeaderResponse#addDependency(String, String,
 	 * String)}, or {@link javax.portlet.HeaderResponse#addDependency(String,
 	 * String, String, String)} are to be referenced in the page's header.
 	 *
-	 * @param portletDependencyJavaScriptEnabled boolean value for whether the
-	 *        JavaScript resource dependencies that are specified in
-	 *        portlet.xml, @{@link javax.portlet.annotations.Dependency},
-	 *        {@link javax.portlet.HeaderResponse#addDependency(String, String,
-	 *        String)}, or {@link javax.portlet.HeaderResponse#addDependency(
-	 *        String, String, String, String)} are to be referenced in the
-	 *        page's header
+	 * @param portletDependencyJavaScriptEnabled whether the JavaScript resource
+	 *        dependencies specified in <code>portlet.xml</code>, @{@link
+	 *        javax.portlet.annotations.Dependency}, {@link
+	 *        javax.portlet.HeaderResponse#addDependency(String, String,
+	 *        String)}, or {@link
+	 *        javax.portlet.HeaderResponse#addDependency(String, String, String,
+	 *        String)} are to be referenced in the page's header
 	 */
 	public void setPortletDependencyJavaScriptEnabled(
 		boolean portletDependencyJavaScriptEnabled) {
@@ -4449,7 +4436,7 @@ public class PortletImpl extends PortletBaseImpl {
 	/**
 	 * The Facebook integration method of the portlet.
 	 *
-	 * @deprecated As of 7.0.0, with no direct replacement
+	 * @deprecated As of Judson (7.1.x), with no direct replacement
 	 */
 	@Deprecated
 	private String _facebookIntegration =
@@ -4608,9 +4595,9 @@ public class PortletImpl extends PortletBaseImpl {
 
 	/**
 	 * <code>True</code> if the portlet's
-	 * serveResource(ResourceRequest,ResourceResponse) method should be invoked
-	 * during a partial action triggered by a different portlet on the same
-	 * portal page.
+	 * <code>serveResource(ResourceRequest,ResourceResponse)</code> method
+	 * should be invoked during a partial action triggered by a different
+	 * portlet on the same portal page.
 	 */
 	private boolean _partialActionServeResource;
 
@@ -4661,8 +4648,8 @@ public class PortletImpl extends PortletBaseImpl {
 	private List<PortletDependency> _portletDependencies;
 
 	/**
-	 * <code>True</code> if the CSS resource dependencies that are specified in
-	 * portlet.xml, @{@link javax.portlet.annotations.Dependency},
+	 * <code>True</code> if the CSS resource dependencies specified in
+	 * <code>portlet.xml</code>, @{@link javax.portlet.annotations.Dependency},
 	 * {@link javax.portlet.HeaderResponse#addDependency(String, String,
 	 * String)}, or {@link javax.portlet.HeaderResponse#addDependency(String,
 	 * String, String, String)} are to be referenced in the page's header.
@@ -4670,8 +4657,8 @@ public class PortletImpl extends PortletBaseImpl {
 	private boolean _portletDependencyCssEnabled;
 
 	/**
-	 * <code>True</code> if the JavaScript resource dependencies that are
-	 * specified in portlet.xml, @{@link javax.portlet.annotations.Dependency},
+	 * <code>True</code> if the JavaScript resource dependencies specified in
+	 * <code>portlet.xml</code>, @{@link javax.portlet.annotations.Dependency},
 	 * {@link javax.portlet.HeaderResponse#addDependency(String, String,
 	 * String)}, or {@link javax.portlet.HeaderResponse#addDependency(String,
 	 * String, String, String)} are to be referenced in the page's header.