+++ portal-impl/src/com/liferay/portal/spring/transaction/DefaultTransactionExecutor.java
@@ -14,9 +14,18 @@
 
 package com.liferay.portal.spring.transaction;
 
+import com.liferay.portal.kernel.dao.orm.EntityCacheUtil;
+import com.liferay.portal.kernel.model.BaseModel;
+import com.liferay.portal.kernel.model.MVCCModel;
 import com.liferay.portal.kernel.transaction.TransactionLifecycleManager;
 
+import java.lang.reflect.Constructor;
+import java.lang.reflect.Field;
+
 import org.aopalliance.intercept.MethodInvocation;
+
+import org.apache.commons.lang.reflect.ConstructorUtils;
+import org.apache.commons.lang.reflect.FieldUtils;
 
 import org.springframework.transaction.PlatformTransactionManager;
 
@@ -63,7 +72,7 @@ public class DefaultTransactionExecutor
 			platformTransactionManager, transactionAttributeAdapter,
 			transactionStatusAdapter);
 
-		return returnValue;
+		return _getUpdatedReturnValue(returnValue);
 	}
 
 	@Override
@@ -148,6 +157,73 @@ public class DefaultTransactionExecutor
 					transactionAttributeAdapter, transactionStatusAdapter);
 			}
 		}
+	}
+
+	private Object _getOneLevelWrapperUpdatedObject(Object returnValue) {
+		Class<?> clazz = returnValue.getClass();
+
+		Field[] fields = clazz.getDeclaredFields();
+
+		if (fields.length != 1) {
+			return returnValue;
+		}
+
+		Field field = fields[0];
+
+		Constructor<?> constructor =
+			ConstructorUtils.getMatchingAccessibleConstructor(
+				clazz, new Class<?>[] {field.getType()});
+
+		try {
+			if (constructor != null) {
+				Object fieldValue = FieldUtils.readDeclaredField(
+					returnValue, field.getName(), true);
+
+				Object newFieldValue = _getUpdatedReturnValue(fieldValue);
+
+				if (newFieldValue != fieldValue) {
+					return constructor.newInstance(fieldValue);
+				}
+			}
+		}
+		catch (Exception e) {
+		}
+
+		return returnValue;
+	}
+
+	private Object _getUpdatedObject(Object object) {
+		if (object instanceof BaseModel<?> && object instanceof MVCCModel) {
+			BaseModel<?> baseModel = (BaseModel<?>)object;
+			MVCCModel mvccModel = (MVCCModel)object;
+
+			Object updatedValue = EntityCacheUtil.getResult(
+				baseModel.isEntityCacheEnabled(), object.getClass(),
+				baseModel.getPrimaryKeyObj());
+
+			if (updatedValue != null) {
+				if (mvccModel.getMvccVersion() <
+						((MVCCModel)updatedValue).getMvccVersion()) {
+
+					return updatedValue;
+				}
+			}
+		}
+
+		return object;
+	}
+
+	private Object _getUpdatedReturnValue(Object returnValue) {
+		if (returnValue instanceof MVCCModel) {
+			if (returnValue instanceof BaseModel<?>) {
+				returnValue = _getUpdatedObject(returnValue);
+			}
+			else {
+				returnValue = _getOneLevelWrapperUpdatedObject(returnValue);
+			}
+		}
+
+		return returnValue;
 	}
 
 }